Fix Health Boost Health Duplication.
This fixed the health duplication for the health boost potion effect.
The health boost potion adds to npc's max health, and will be returned by Player.getMaxHealth().
Once we apply this to the new player with Player.setMaxHealth(), it becomes permanent.
Players can reapatedly abuse this to gain limitless permenant health boosts.

Merges #107